implemented dashbaord

+++ routes/blog.route.js
@@ -5,6 +5,7 @@ const User = require("../models/user.schema");
 const Comment = require("../models/comment.schema");
 const Blog = require("../models/blog.schema");
 const {restrictToLoggedinUserOnly} = require("../middlewares/authentication.middleware");
+const { error } = require("console");
 
 const router = Router();
 
@@ -20,6 +21,7 @@ const storage = multer.diskStorage({
 
 const upload = multer({storage : storage}); 
 
+// Homepage 
 router.get('/',async (req,res) => {
     const allBlogs = await Blog.find({status: "PUBLISHED"});
     return res.render('home',{
@@ -28,24 +30,41 @@ router.get('/',async (req,res) => {
     });
 });
 
-router.get('/blog/:id',async (req,res) => {
-    const blog = await Blog.findById(req.params.id);
-    await Blog.findByIdAndUpdate(req.params.id,{
-        viewsCount : blog.viewsCount+=1
-    });
-    const author = await User.findById(blog.createdBy);
-    const comments = await Comment.find({ blogId: blog._id })
-    .populate('userId')
-    .sort({ createdAt: -1 }); // Sort comments in descending order by createdAt timestamp
-
-    return res.render('viewBlog',{
-        user: req.user,
-        blog: blog,
-        author:author,
-        comments : comments,
-    });
+// display logges user blogs
+router.get('/publishedBlog',async (req,res) => {
+    const blogs = await Blog.find({createdBy:req.user._id});
+    
+    return res.render("publishedBlogs",{blogs:blogs,user: req.user,});
 });
 
+// View Particular blog
+router.get('/blog/:id',async (req,res) => {
+    try{
+        const blog = await Blog.findById(req.params.id);
+        if(blog.status == "DRAFT" && blog.createdBy != req.user._id) throw error;
+        if(blog.status == "PUBLISHED"){
+            await Blog.findByIdAndUpdate(req.params.id,{
+                viewsCount : blog.viewsCount+=1
+            });
+        }
+        const author = await User.findById(blog.createdBy);
+        const comments = await Comment.find({ blogId: blog._id })
+        .populate('userId')
+        .sort({ createdAt: -1 }); // Sort comments in descending order by createdAt timestamp
+    
+        return res.render('viewBlog',{
+            user: req.user,
+            blog: blog,
+            author:author,
+            comments : comments,
+        });
+    }
+    catch(error){
+        return res.render("404",{user:req.user});
+    }
+});
+
+// create  a blog
 router.get('/addBlog',restrictToLoggedinUserOnly,(req,res) => {
     return res.render("addBlog",{user: req.user,});
 }); 
@@ -63,6 +82,8 @@ router.post('/addBlog',restrictToLoggedinUserOnly,upload.single('coverImage'),as
     return res.redirect(`blog/${blog._id}`);
 });
 
+
+// draft blog 
 router.post('/draftBlog',restrictToLoggedinUserOnly,upload.single('coverImage'),async (req,res) => {
     const {title,content,description} = req.body;
     const blog  = await Blog.create({
@@ -76,38 +97,39 @@ router.post('/draftBlog',restrictToLoggedinUserOnly,upload.single('coverImage'),
     return res.redirect(`/`);
 });
 
-router.get('/publishedBlog',async (req,res) => {
-    const blogs = await Blog.find({createdBy:req.user._id});
-    return res.render("publishedBlogs",{blogs:blogs,user: req.user,});
+
+// comment 
+router.post('/comment/:id',restrictToLoggedinUserOnly,async(req,res)=>{
+    try{
+        const blogId = req.params.id;
+        const blog = await Blog.findById(req.params.id);
+        if(blog.createdBy != req.user._id && blog.status == "DRAFT") throw error;
+        const {comment} = req.body;
+        await Comment.create({
+            body : comment,
+            userId: req.user._id,
+            blogId : blogId
+        });
+        return res.redirect(`/blog/${blogId}`);
+    }
+    catch(error){
+        return res.render("404",{user:req.user});
+    }
 });
 
-router.post('/comment/:id',restrictToLoggedinUserOnly,async(req,res)=>{
-    const blogId = req.params.id;
-    const {comment} = req.body;
-    await Comment.create({
-        body : comment,
-        userId: req.user._id,
-        blogId : blogId
-    });
-    return res.redirect(`/blog/${blogId}`);
-});
-
-router.post('/comment/:id',restrictToLoggedinUserOnly,async(req,res)=>{
-    const blogId = req.params.id;
-    const {comment} = req.body;
-    await Comment.create({
-        body : comment,
-        userId: req.user._id,
-        blogId : blogId
-    });
-    return res.redirect(`/blog/${blogId}`);
-});
 
 router.get('/editBlog/:id',restrictToLoggedinUserOnly,async (req,res) => {
-    const blog = await Blog.findById(req.params.id);
-    return res.render("editBlog",{blog:blog,user: req.user});
+    try{
+        const blog = await Blog.findById(req.params.id);
+        if(blog.createdBy != req.user._id) throw error;
+        return res.render("editBlog",{blog:blog,user: req.user});
+    }catch(error){
+        return res.render("404",{user:req.user});
+    }
 });
 
+
+// edit blog
 router.post('/editBlog/:id',restrictToLoggedinUserOnly,upload.single('coverImage'),async (req,res) => {
     const {title,body,description ,status} = req.body;
 
@@ -126,9 +148,18 @@ router.post('/editBlog/:id',restrictToLoggedinUserOnly,upload.single('coverImage
     return res.redirect(`/publishedBlog`);
 });
 
+
+//delete blog
 router.get('/deleteBlog/:id',async (req,res) => {
-    await Blog.findByIdAndDelete(req.params.id);
-    return res.redirect(`/publishedBlog`);
+    try{
+        const blog = await Blog.findById(req.params.id);
+        if(blog.createdBy != req.user._id) throw error;
+        await Blog.findByIdAndDelete(req.params.id);
+        return res.redirect(`/publishedBlog`);
+    }
+    catch(error){
+        return res.render("404",{user:req.user});
+    }
 });
 
 // Like a blog
